Squashing the latest 4 commits. kind of like PR Updated source directory.
Updated the genparticle class to have a layer option, updated the sampling section to include a sensitive layer counter for hadrons, muons, and neutrons

PFCAL was changed to have a distinct tracks vector and gen particle vector.  The tracks vector stores the first instance of all long lasting gen hadrons with > 10 MeV.  The EventAction class was changed to track the number of hadrons, muons and neutrons in each sensitive layer.  The energy weighted RMS centroid distance was corrected to only be measured in transverse plane

stop lsf spam